Removed server.js
